refactor: Pawn 클래스 Piece 클래스로 상위 개념으로 리팩토링
