+++ src/main/java/softeer2nd/domain/VO/Type.java
@@ -1,0 +1,10 @@
+package softeer2nd.domain.VO;
+
+public enum Type {
+    PAWN,
+    KNIGHT,
+    ROOK,
+    BISHOP,
+    QUEEN,
+    KING,
+}
